marine-draft/requirements.txt
streamlit
reportlab
pillow
matplotlib
